Redo the JavaScript minification
This commit rewrites how we do the JavaScript minification to simplify the process
by eliminating the combination step and by assuming there is only ever one JavaScript
file to minify. This ensure the script only works in the most basic circumstances
and therefore avoids making assumptions about what to do when more complicated changes
are made (such as multiple JavaScript source files are present and are references from
the main html file).