Update readme to include information about strings
The branding pack provides the functionality to override the default strings used in the HTML reader, however up until this point a current list of the default strings in use by the HTML reader was not provided to clients easily and therefore knowledge of what could be overridden and how was lacking. This commit fixes this issue.