Add Jekyll metadata
This commit adds some metadata to the Jekyll configuration file. It adds a title and description to the site. It also toggles a flag in the custom theme that adds download buttons to the page headers. This latter should make it simple for clients to download bundles of the latest branding pack straight from the documentation pages.