Split the readme
This commit splits the readme into separate files for each of its sections. Some sections are in their own subfolders to better group content. All of it is now within the `/docs` subfolder. Note that no links or context have yet been changed: this commit simply splits the content up.

In its place, the base readme now contains some filler text with links to the company homepage and to the table of contents. This should be updated with some more useful information when possible.