Update title levels and links
This commit reduces all the title levels in the various markdown files. Since they are no longer part of a single large file, each can have its own primary header, and other titles cascade down from that. For the API pages, this means a reduction of two title levels across the board, and for everything else a reduction of one.

Also in this commit are updates to all the links in the table of contents page. This is more essential than ever, so it is critical that the links function correctly.