Tweak the default layout
This commit makes some minor modifications to the default layout. The first change is to add a "return to the top" link to the page footer. This should make scrolling up simpler on long pages. The second, and more important change, is to add a link opposite the download buttons. It returns viewers to the table of contents page, which is used a bit like an index for now. Note that the description / subtitle has also had a link injected to the company page, which will be present on all site pages.