Fix the contents interface formatting
This commit separates the example JSON with sample contents data from the surrounding list. The Jekyll parser was breaking the multi-line code snippet when it was kept as part of the list. Since the list is simple, this is currently trivial to do. It also makes it possible to add JSON syntax highlighting to the snippet.