Move the contents back to the main page
This commit moves the table of contents back to the main readme. This required updating all the relative links again. But since there was little information for the opening page that was not better suited to the introduction page, it seems better to keep the contents on the homepage.

Note that since the home page is now effectively the contents page, the opening sentences have been reworded. They are more basic, and do not attempt to convey anything other than a summary of the project. All other information is deferred to the sub pages.