Remove the max-width on content
This commit removes the max-width limit on primary content. It was limited to 640px by the theme. This commit instead adds explicit margins of 100px for screens wider than 840px. Screens between 640 and 840 pixels wide will now interpolate margins, meaning that for those 200 pixels, the margins will increase instead of the content size.