Adds a copy task
This commit adds a new task to the build script. It copies all of the built files into their own dedicated folder. This allows a single task to build, copy and then clean all files needed for the pack. This should also make it easier to determine which files are necessary for the pack, and which are simply present due to git, the build process, and hosted documentation.