Remove redundant classes from info pages
This commit removes some classes from the checkbox label. Two of the three removed are for the `noTouch` flag that tells the Reader not to dismiss the page when interacted with. In this instance, the classes are unneeded because the elements are nested within another with the class. That means that events from the inner element must bubble up to the outer element with the class, before they reach the document body (which is where the dismissal is triggered). Thus it is safe to remove the inner class.

The other class removed here is an instance of `yudu_localisable`. This class typcially informs the Reader that the element has text in need ot localising. In this instance however, the element with localisable text is nested within the element with the class, and itself has the class anyway. Thus once again, the class is redundant.