Use HammerJS for the sharing popup
This commit switches the sharing popup from using native events to using HammerJS. The whole popup is now assigned a single Hammer Manager, in the style of the toolbar. Callbacks for the various buttons are keyed to the trigger element ID. When the element is tapped, the handler is fired, which checks the target ID, and looks for a callback keyed by that ID. Unlike the toolbar handler, the sharing handler passes the event on to the triggered callback, so it can make use of it if need be.

Note that to ensure the share-type buttons trigger correctly, the internal elements have been given the CSS property `touch-action: none` to ensure pointer events pass through them to the container with the expected ID. This has been done by adding a new class to the CSS stylesheet imported by the toolbar, and by adding that class to the hardcoded inner object, and dynamically to the images.

In addition to both of the above, some of the button creation code has been extracted to a common helper. Since each of the buttons was performing identical actions, a shared function for the three seemed ideal.