Extract duplicate code from sharing callbacks
This commit builds on the refactoring in the parent commit. As before, some repeated logic has been encapsulated in a helper function. This helper takes the specific callback that should be attributed to a sharing button, and adds the shared UI managment functionality. A sanity check for the function is added to ensure a valid object has been made available from the core, in case the default callbacks are used.