Convert the contents popup to HammerJS
This commit converts the contents menu popup to use HammerJS instead of native events. Like the other conversions, a single manager has been added for the entire list. On activation, it checks for a known ID in a hash of callbacks. If a valid function is found, it is triggered, and the event is passed on.

Unlike the previous conversions, this one puts the ID into the contents link dataset. The items do not currently have IDs, so utilising the dataset allows a non-clicting way to add IDs that only have to be unique among the content items themselves. For now, a self-contained generator that simply counts up is being used to generate these unique IDs.