Prepare for converting the thumbnail bar
This commit makes some preparatory changes for converting the thumbnails bar to HammerJS. The thumbnails bar will likely not be converted fully. Like the other brandable submodules, only those elements that can be tapped are intended for inclusion in this initial conversion work. Nevertheless, even that requires a new tap manager, and a mechanic for registering callbacks. This commit uses an implementation similar to the contents popup - namespace-unique IDs are associated with an element's dataset ID variable. This keeps HTML element IDs clear when not already in use, keeping the global namespace clearer. It further permits for simple identification between the various subsets of objects within the thumbnails script.