Use loading flags as well as events
This commit updates the default loading script. In addition to subscribing to events from the core, the script's initial state is received from glags set to the API settings object. This allows handling for cases when loading progress is made - or even completed - before the script is ready to handle events.