Extract a method and pass through orientation
This commit extracts some logic from inside various loops. The logic was calculating the rescaled size for an icon, and applying it for the current orientation. To make the newly extracted method even more useful, the orientation has also been made into a parameter, to allow sizes to be set for both orientations instead of requiring it be for the current one only. This has required an inner method also accept the orientation as a parameter, which in turn meant extracting it to another call. However, this makes it more explicit what is actually being set in these locations, so seems an advantage overall.