Update an icon's size on page decryption
This commit ensures that when a page has its data decrypted, its sizes will also be updated. This could be relevant if the encrypted and decrypted pages had different aspect ratios - which is likely since the default encrypted page is more square than more. Note that the icon's sizes for both orientations must be considered, since there is currently no way to mark a size as requiring updating, however the orientation is only updated if it has been globally initialised. A global initialisation post-decryption should detect the correct sizes in any event.